Add ignore bg color for header text color
In macOS with the newer python 3.10.x that has working dark mode - the column headers get their background color from the background of the treeview.  In a recent commit, support was added for determining if the header text should be black or white based on the luminance of the treeview background.  This does not work with older python on macOS where full dark mode is not supported, so a checkbox is added to ignore the treeview background color when calculating the header text color.  Default is black - but the invert checkbox still works, and can set it to white if needed.