Add should_set_header_text() check
On macOS, with tk versions that support full dark mode, the column headers take the background color only in dark mode - falling back to white in light mode regardless of the background color.  This can cause the headers to be unreadable at times.

This function returns whether or not we should be adjusting the header text color based on the background color - which is always True on non-macOS systems, and True on macOS when in dark mode.