Use tk call to ask the window direct if dark
Instead of asking macOS if we're in dark mode and then hoping for support in the window - we now use a tk call to check the window directly.  If this is not supported in the current tk version, we assume light mode which should prevent erroneously adjusting the header text colors.