Rework some dragging and pasting behavior
Prior, empty collections (arrays, dictionaries) were always treated as closed - so all drags and pastes treated them as non-collections and always drug/pasted as a sibling.  This commit treats them as open, and corrects some dragging behavior that doesn't move a node to the top level if drug to the bottom of the tree view.